activity_regis_02/backend_act/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
blessings==1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
curtsies==0.3.10
cwcwidth==0.1.6
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.48.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
jedi==0.18.1
oauthlib==3.2.0
parso==0.8.3
Pillow==9.0.0
prompt-toolkit==3.0.28
ptpython==3.0.20
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.27
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
